+++ changelog.md
@@ -11,6 +11,8 @@ Codion Change Log
 - Conditional no longer implements Consumer.
 - Value.Change renamed ValueChange and moved up to package level, factory method added.
 - DefaultValueChange moved up to package level, implements equals() and hashCode().
+### is.codion.common.utilities
+- Exceptions.runtime() resulting RuntimeException now uses the stacktrace from the cause.
 ### is.codion.common.model
 - DefaultConditionModel, caseSensitive now included in locked state.
 ### is.codion.swing.common.model